+++ lib/src/common/constants.dart
@@ -1,2 +1,0 @@
-double kLargeScreenMinWidth = 1100;
-double kSmallScreenMaxWidth = 600;