+++ lib/src/domain/enums/responsive_type.dart
@@ -1,0 +1,14 @@
+/// Responsive with all screen types.
+enum ResponsiveType {
+  /// The screen is desktop size.
+  desktop,
+
+  /// The screen is phone size.
+  phone,
+
+  /// The screen is tablet size.
+  tablet,
+
+  /// The screen is watch size.
+  watch,
+}